resolve ruby 2.6 deprecation of BigDecimal.new
previously, NCore::BigMoney was a subclass of BigDecimal. in ruby 2.6,
BigDecimal.new is deprecated and prints a warning every use. this changes
BigMoney to be a delegator to BigDecimal instead.

BigMoney's original goal of modifying #to_s to print to 2 decimals by default
is maintained. just about any other operation on BigMoney will cause it to turn
back into a standard BigDecimal.