Preserve 'metadata' and 'errors' keys as attributes when inside a standard data payload
Previously these attribute keys were improperly stored using the global `metadata=` and `errors=` accessors, which could interfere with the normal use of `metadata` and `errors`.

Now these attribute keys are kept as part of the attributes store. Because of the non-attribute versions of `metadata` and `errors`, these attribute keys must be accessed using one of:

1) Use hash-like syntax:  `obj[:metadata]` or `obj[:errors]`

2) Create an alternate named reader that uses the above:
    def resource_metadata
      self[:metadata]
    end